Text Scrambler for DnD clues.
This python code takes a .txt file, and returns it altered. This is done by either:
'smudging' the characters (replacing a fraction of them =to an input value between 0 and 1, with punctuation of similar size). Default 0.3
'Burning Holes' Making circles of black squares in the text file, supply radius and number of holes. Default 2 holes of radius 5 characters.
'Charring the Edges'  Black squares encroaching from the boarder.  Default 2 +/- 1. 

It takes less than a second. and makes a new file so the original isn't lost, which you can then forward to players as the .txt or copy the text yourself and paste it in a chat.  

I use this to alter text information made for dnd. i.e the papers burn during the fight, if the players take longer to extinguish the flames, the parameters of the code are increased for burning effects. Or if a book is thrown into water, smudging increases. etc.